Improve uninstall statement

+++ uninstall.php
@@ -5,11 +5,15 @@
     }
 
     // drop the database table
-    global $wpdb;
-    $wpdb->query("DROP TABLE IF EXISTS {$wpdb->prefix}cities");
+    if ( false == get_option( 'acfcs_preserve_settings' ) ) {
+	    global $wpdb;
+	    $wpdb->query("DROP TABLE IF EXISTS {$wpdb->prefix}cities");
 
-	$target_folder = wp_upload_dir()['basedir'] . '/acfcs';
-    rmdir( $target_folder );
+	    $target_folder = wp_upload_dir()['basedir'] . '/acfcs';
+	    rmdir( $target_folder );
+
+	    delete_option( 'acfcs_preserve_settings' );
+    }
 
     // drop all settings from db (if any)
 ?>
